mentor_mentee/backend/requirements.txt
# Core dependencies
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3

# API Documentation
drf-yasg==1.21.7
setuptools>=68.0.0

# Utils
python-dotenv==1.0.0
dj-database-url==2.1.0

# Optional database adapters - uncomment if needed
mysqlclient==2.2.0
# psycopg2-binary==2.9.7 